Break apart safir.database for ease of maintenance
Convert safir.database from a single-file module to a directory
and separate the code into internal modules for ease of maintenance.
Upcoming PRs will be adding more code to safir.database, and this
will avoid dumping all of that code into a single large file.
